requirements_linux_venv.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi-bindings==21.2.0
argon2-cffi==23.1.0
arrow==1.3.0
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.0.4
attrs==25.1.0
babel==2.18.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2026.2.25
cffi==1.16.0
charset-normalizer==3.3.2
cinemagoer==2023.5.1
click==8.3.1
cloudpathlib==0.16.0
cloudpickle==3.1.2
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.0
cycler==0.12.1
cymem==2.0.13
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
executing==2.2.1
fastjsonschema==2.19.1
filelock==3.25.2
Flask==3.0.2
flatbuffers==25.12.19
fonttools==4.51.0
fqdn==1.5.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.3.0
future==1.0.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
greenlet==3.3.2
grpcio==1.80.0
gunicorn==21.2.0
h11==0.14.0
h2==4.3.0
h5py==3.16.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
hyperframe==6.1.0
hyperopt==0.2.7
idna==3.6
IMDbPY==2022.7.9
ipykernel==6.29.4
ipython==8.22.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema-specifications==2023.12.1
jsonschema==4.21.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter==1.0.0
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.4
keras==3.13.2
kiwisolver==1.4.5
langchain-community==0.0.38
langchain-core==0.1.53
langchain-text-splitters==0.0.2
langchain==0.1.16
langcodes==3.5.1
langgraph==0.0.40
langsmith==0.1.147
libclang==18.1.1
lxml==6.0.2
markdown-it-py==4.0.0
Markdown==3.10.2
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.1.7
matplotlib==3.8.4
mdurl==0.1.2
mistune==3.2.0
ml-dtypes==0.3.2
mmh3==5.2.1
mpmath==1.3.0
msgpack==1.0.8
multidict==6.7.1
multitasking==0.0.12
murmurhash==1.0.15
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.18.1
nbclient==0.9.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.8.1
notebook==7.1.2
notebook_shim==0.2.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
optree==0.19.0
opt_einsum==3.4.0
orjson==3.11.8
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.6
peewee==4.0.3
permetrics==2.0.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
plotly==5.20.0
postgrest==2.28.3
preshed==3.0.13
prometheus_client==0.24.1
prompt-toolkit==3.0.43
propcache==0.4.1
protobuf==4.25.9
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.17.2
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.1.2
pyroaring==1.0.4
python-dateutil==2.9.0
python-dotenv==1.0.1
python-json-logger==4.1.0
pytz==2026.1.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.7.2
QtPy==2.4.3
ray==2.49.2
realtime==2.28.3
referencing==0.37.0
regex==2026.3.32
requests-toolbelt==1.0.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.4.2
scipy==1.12.0
seaborn==0.13.2
Send2Trash==2.1.0
setuptools==82.0.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy==3.7.4
SQLAlchemy==2.0.29
srsly==2.5.3
stack-data==0.6.3
storage3==2.28.3
streamlit==1.33.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase-auth==2.28.3
supabase-functions==2.28.3
supabase==2.28.3
sympy==1.14.0
tenacity==8.5.0
tensorboard-data-server==0.7.2
tensorboard==2.16.2
tensorflow==2.16.1
termcolor==3.3.0
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.6.0
tinycss2==1.5.1
tokenizers==0.15.2
toml==0.10.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.39.3
typer==0.9.4
types-python-dateutil==2.9.0.20260323
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.2.1
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.6.0
weasel==0.3.4
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.7
wheel==0.45.1
widgetsnbextension==4.0.15
wrapt==2.1.2
yarl==1.23.0
yfinance==0.2.37
zstandard==0.25.0